+++ tests/test_security.py
@@ -12,50 +12,60 @@ class TestSecurityCSP(unittest.TestCase):
 
     def test_init(self):
         generator = CspGenerator()
-        self.assertEqual(generator.get_csp(), "default-src 'self'; object-src 'none';")
+        self.assertEqual(generator.to_string(), "default-src 'self'; object-src 'none';")
+
+    def test_init_str_output(self):
+        generator = CspGenerator()
+        self.assertEqual(str(generator), "default-src 'self'; object-src 'none';")
+
+    def test_init_dict_output(self):
+        generator = CspGenerator()
+        self.assertEqual(
+            generator.to_dict(), {"default-src": ["'self'"], "object-src": ["'none'"]}
+        )
 
     def test_init_none(self):
         generator = CspGenerator(CspGenerator.NONE)
-        self.assertEqual(generator.get_csp(), "default-src 'none'; object-src 'none';")
+        self.assertEqual(generator.to_string(), "default-src 'none'; object-src 'none';")
 
     def test_init_list(self):
         generator = CspGenerator([CspGenerator.NONE, self.test_domain])
         self.assertEqual(
-            generator.get_csp(),
+            generator.to_string(),
             f"default-src 'none' {self.test_domain}; object-src 'none';",
         )
 
     def test_init_empty_string(self):
         generator = CspGenerator("")
-        self.assertEqual(generator.get_csp(), "default-src 'self'; object-src 'none';")
+        self.assertEqual(generator.to_string(), "default-src 'self'; object-src 'none';")
 
     def test_init_list_of_empty_strings(self):
         generator = CspGenerator([""])
-        self.assertEqual(generator.get_csp(), "default-src 'self'; object-src 'none';")
+        self.assertEqual(generator.to_string(), "default-src 'self'; object-src 'none';")
 
     def test_init_allow_objects(self):
         generator = CspGenerator([""], allow_objects=True)
-        self.assertEqual(generator.get_csp(), "default-src 'self';")
+        self.assertEqual(generator.to_string(), "default-src 'self';")
 
     def test_add_directive(self):
         generator = CspGenerator()
         generator.script_src(self.test_domain)
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn(f"script-src 'self' {self.test_domain};", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn(f"script-src 'self' {self.test_domain};", generator.to_string())
 
     def test_add_directive_none(self):
         generator = CspGenerator()
         generator.script_src(CspGenerator.NONE)
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn("script-src 'none';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn("script-src 'none';", generator.to_string())
 
     def test_add_directive_multiple(self):
         generator = CspGenerator()
         generator.script_src(self.test_domain, self.test_domain_2)
-        self.assertIn("default-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
         self.assertIn(
             f"script-src 'self' {self.test_domain} {self.test_domain_2};",
-            generator.get_csp(),
+            generator.to_string(),
         )
 
     def test_add_directive_duplicates(self):
@@ -63,19 +73,19 @@ class TestSecurityCSP(unittest.TestCase):
         generator.script_src(
             self.test_domain, self.test_domain_2, self.test_domain, self.test_domain
         )
-        self.assertIn("default-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
         self.assertIn(
             f"script-src 'self' {self.test_domain} {self.test_domain_2};",
-            generator.get_csp(),
+            generator.to_string(),
         )
 
     def test_add_directive_list(self):
         generator = CspGenerator()
         generator.script_src([self.test_domain, self.test_domain_2])
-        self.assertIn("default-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
         self.assertIn(
             f"script-src 'self' {self.test_domain} {self.test_domain_2};",
-            generator.get_csp(),
+            generator.to_string(),
         )
 
     def test_add_directive_mixed_types(self):
@@ -89,73 +99,73 @@ class TestSecurityCSP(unittest.TestCase):
             f"{self.test_domain_2} {test_domain_3}",
             [test_domain_4, f"{test_domain_5} {test_domain_6}"],
         )
-        self.assertIn("default-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
         self.assertIn(
             f"script-src 'self' {self.test_domain} {self.test_domain_2} {test_domain_3} {test_domain_4} {test_domain_5} {test_domain_6};",
-            generator.get_csp(),
+            generator.to_string(),
         )
 
     def test_add_directive_space_separated_list(self):
         generator = CspGenerator()
         generator.script_src(f"{self.test_domain} {self.test_domain_2}")
-        self.assertIn("default-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
         self.assertIn(
             f"script-src 'self' {self.test_domain} {self.test_domain_2};",
-            generator.get_csp(),
+            generator.to_string(),
         )
 
     def test_add_directive_empty(self):
         generator = CspGenerator()
         generator.script_src()
-        self.assertIn("default-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
         self.assertNotIn(
             f"script-src 'self' {self.test_domain} {self.test_domain_2};",
-            generator.get_csp(),
+            generator.to_string(),
         )
 
     def test_add_directive_duplicated_self(self):
         generator = CspGenerator()
         generator.script_src(CspGenerator.SELF)
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertNotIn("script-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertNotIn("script-src 'self';", generator.to_string())
 
     def test_add_directive_existing_self(self):
         generator = CspGenerator()
         generator.script_src([CspGenerator.SELF, self.test_domain])
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn(f"script-src 'self' {self.test_domain};", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn(f"script-src 'self' {self.test_domain};", generator.to_string())
 
     def test_add_directive_existing_none(self):
         generator = CspGenerator()
         generator.script_src([CspGenerator.NONE, self.test_domain])
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn(f"script-src 'none' {self.test_domain};", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn(f"script-src 'none' {self.test_domain};", generator.to_string())
 
     def test_add_directive_omit_self(self):
         generator = CspGenerator()
         generator.script_src(self.test_domain, omit_self=True)
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn(f"script-src {self.test_domain};", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn(f"script-src {self.test_domain};", generator.to_string())
 
     def test_add_directive_list_omit_self(self):
         generator = CspGenerator()
         generator.script_src([self.test_domain, self.test_domain_2], omit_self=True)
-        self.assertIn("default-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
         self.assertIn(
-            f"script-src {self.test_domain} {self.test_domain_2};", generator.get_csp()
+            f"script-src {self.test_domain} {self.test_domain_2};", generator.to_string()
         )
 
     def test_add_directive_omit_self_existing_none(self):
         generator = CspGenerator()
         generator.script_src([CspGenerator.NONE, self.test_domain], omit_self=True)
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn(f"script-src 'none' {self.test_domain};", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn(f"script-src 'none' {self.test_domain};", generator.to_string())
 
     def test_add_disallow_directive(self):
         generator = CspGenerator()
         generator.disallow("script-src")
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn("script-src 'none';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn("script-src 'none';", generator.to_string())
 
     def test_add_directive_chained(self):
         generator = CspGenerator()
@@ -165,33 +175,33 @@ class TestSecurityCSP(unittest.TestCase):
         generator = CspGenerator()
         generator.script_src(self.test_domain)
         generator.style_src(self.test_domain_2)
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn(f"script-src 'self' {self.test_domain};", generator.get_csp())
-        self.assertIn(f"style-src 'self' {self.test_domain_2};", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn(f"script-src 'self' {self.test_domain};", generator.to_string())
+        self.assertIn(f"style-src 'self' {self.test_domain_2};", generator.to_string())
 
     def test_add_multiple_same_directives(self):
         generator = CspGenerator()
         generator.script_src(self.test_domain)
         generator.script_src(self.test_domain_2)
-        self.assertIn("default-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
         self.assertIn(
             f"script-src 'self' {self.test_domain} {self.test_domain_2};",
-            generator.get_csp(),
+            generator.to_string(),
         )
 
     def test_add_multiple_same_directives_replace(self):
         generator = CspGenerator()
         generator.script_src(self.test_domain)
         generator.script_src(self.test_domain_2, replace=True)
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn(f"script-src 'self' {self.test_domain_2};", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn(f"script-src 'self' {self.test_domain_2};", generator.to_string())
 
     def test_add_custom_directive(self):
         generator = CspGenerator()
         generator.custom_src("custom-directive", self.test_domain)
-        self.assertIn("default-src 'self';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
         self.assertIn(
-            f"custom-directive 'self' {self.test_domain};", generator.get_csp()
+            f"custom-directive 'self' {self.test_domain};", generator.to_string()
         )
 
     def test_add_directive_sources(self):
@@ -221,60 +231,60 @@ class TestSecurityCSP(unittest.TestCase):
         for directive, method in directives:
             generator = CspGenerator()
             getattr(generator, method)(self.test_domain)
-            self.assertIn("default-src 'self';", generator.get_csp())
+            self.assertIn("default-src 'self';", generator.to_string())
             self.assertIn(
-                f"{directive} 'self' {self.test_domain};", generator.get_csp()
+                f"{directive} 'self' {self.test_domain};", generator.to_string()
             )
 
     def test_add_report_uri(self):
         generator = CspGenerator()
         report_uri = "https://report.example.com"
         generator.report_uri(report_uri)
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn(f"report-uri {report_uri};", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn(f"report-uri {report_uri};", generator.to_string())
 
     def test_add_empty_report_uri(self):
         generator = CspGenerator()
         generator.report_uri("")
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertNotIn("report-uri", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertNotIn("report-uri", generator.to_string())
 
     def test_add_report_to(self):
         generator = CspGenerator()
         report_endpoint_name = "csp_report_endpoint"
         generator.report_to(report_endpoint_name)
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn(f"report-to {report_endpoint_name};", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn(f"report-to {report_endpoint_name};", generator.to_string())
 
     def test_add_empty_report_to(self):
         generator = CspGenerator()
         generator.report_to("")
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertNotIn("report-to", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertNotIn("report-to", generator.to_string())
 
     def test_add_require_trusted_types_for(self):
         generator = CspGenerator()
         generator.require_trusted_types_for()
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn("require-trusted-types-for 'script';", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn("require-trusted-types-for 'script';", generator.to_string())
 
     def test_add_sandbox(self):
         generator = CspGenerator()
         generator.sandbox()
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn("sandbox;", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn("sandbox;", generator.to_string())
 
     def test_add_sandbox_value(self):
         generator = CspGenerator()
         generator.sandbox("allow-scripts")
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn("sandbox allow-scripts;", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn("sandbox allow-scripts;", generator.to_string())
 
     def test_add_sandbox_invalid_value(self):
         generator = CspGenerator()
         generator.sandbox("pizza")
-        self.assertIn("default-src 'self';", generator.get_csp())
-        self.assertIn("sandbox;", generator.get_csp())
+        self.assertIn("default-src 'self';", generator.to_string())
+        self.assertIn("sandbox;", generator.to_string())
 
 
 class TestCommonSecurityHeaders(unittest.TestCase):